changed the cell returned type to double instead of int in order to show the small variations in the sensitivity analysis

+++ src/main/java/uk/ac/bcu/kbe/bioen/dao/ExcelReader.java
@@ -67,9 +67,9 @@ public class ExcelReader {
         return  getValue(getCell(sheetIndex,rowIndex,columnIndex));
     }
 
-    public int getCellValueNumber(int sheetIndex, String cellId) {
+    public double getCellValueNumber(int sheetIndex, String cellId) {
         double doubleValue = Double.parseDouble(getValue(getCell(sheetIndex, cellId)));
-        return (int) doubleValue;
+        return  doubleValue;
     }
 
     public String getCellValueString(int sheetIndex, String cellId) {
